plot_scripts/plot_comparison.py
import sys
import pandas as pd
import os
import matplotlib.pyplot as plt
import seaborn as sns
import pathlib as pl
import numpy as np
import yaml

sys.path.append(str(list(pl.Path(__file__).parents)[1]))
os.chdir(str(list(pl.Path(__file__).parents)[1]))

from modules.plotting.compare_MCS_deter import *
from modules.plotting.MCS_metric_deterministic_reader import *

plot_pdf_cdf()
compare_weights()
pie_chart_energy()
